idea-hn-47354996-show-hn--ava---ai-voice-agent-for-tradit/tests/test_call.py
import pytest
import httpx
import asyncio

# This is a very basic placeholder.
# A real test would involve:
# 1. Mocking ARI client and AI providers.
# 2. Using a test client for FastAPI.
# 3. Potentially simulating Asterisk calls.

@pytest.mark.asyncio
async def test_health_endpoint_external():
    """
    Tests the health endpoint of the running FastAPI application.
    This test requires the Docker containers to be up and running.
    """
    async with httpx.AsyncClient(base_url="http://localhost:8000") as client:
        try:
            response = await client.get("/api/health", timeout=5)
            assert response.status_code == 200
            assert "status" in response.json()
            assert "active_calls" in response.json()
            assert "asterisk_connected" in response.json()
            # We expect asterisk_connected to be true if Asterisk is up
            # and AVA successfully connected.
            assert response.json()["asterisk_connected"] is True
        except httpx.ConnectError:
            pytest.fail("FastAPI app not running or not reachable at http://localhost:8000")
        except Exception as e:
            pytest.fail(f"An unexpected error occurred: {e}")

# More tests would go here, e.g., for config updates, call logs, etc.
# For ARI interaction, extensive mocking would be required.
# Example of a simple config test (requires app to be running)
@pytest.mark.asyncio
async def test_config_update_external():
    async with httpx.AsyncClient(base_url="http://localhost:8000") as client:
        try:
            # Get initial config
            initial_config_response = await client.get("/api/config")
            assert initial_config_response.status_code == 200
            initial_prompt = initial_config_response.json()["system_prompt"]

            # Update config
            new_prompt = "You are a friendly robot assistant."
            update_response = await client.post("/api/config", json={"system_prompt": new_prompt})
            assert update_response.status_code == 200
            assert update_response.json()["status"] == "updated"

            # Get updated config
            updated_config_response = await client.get("/api/config")
            assert updated_config_response.status_code == 200
            assert updated_config_response.json()["system_prompt"] == new_prompt

            # Revert to initial prompt for clean up
            await client.post("/api/config", json={"system_prompt": initial_prompt})

        except httpx.ConnectError:
            pytest.skip("FastAPI app not running, skipping external config test.")
        except Exception as e:
            pytest.fail(f"An unexpected error occurred during config test: {e}")
