idea-hn-47354996-show-hn--ava---ai-voice-agent-for-tradit/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ari==0.2.1
pydub==0.25.1
openai==1.12.0
pyyaml==6.0.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
websockets==12.0
httpx==0.26.0
aiohttp==3.9.0
